Split UI variant helpers
