+++ frontend/components/ui/badge.tsx
@@ -43,4 +43,4 @@ function Badge({
   )
 }
 
-export { Badge, badgeVariants }
+export { Badge }
