+++ frontend/components/ui/toggle-group.tsx
@@ -4,7 +4,7 @@ import * as ToggleGroupPrimitive from '@radix-ui/react-toggle-group'
 import { type VariantProps } from 'class-variance-authority'
 import * as React from 'react'
 
-import { toggleVariants } from '@/components/ui/toggle'
+import { toggleVariants } from '@/components/ui/toggle-variants'
 import { cn } from '@/lib/utils'
 
 const ToggleGroupContext = React.createContext<
